# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Seismic, a developer time tracking platform.",
        "title": "Seismic API",
        "contact": {
            "email": "hello@seismic.icu"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/filters/languages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filters"
                ],
                "summary": "Get distinct languages coded",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/filters/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filters"
                ],
                "summary": "Get distinct projects coded",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helpers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}